requirements.txt
# 飞书 SDK
lark-oapi>=1.2.0

# 异步支持
aiohttp>=3.9.0

# 配置管理
pyyaml>=6.0
python-dotenv>=1.0.0

# 安全
cryptography>=41.0.0

# 日志
loguru>=0.7.0

# 类型检查
pydantic>=2.0.0
